feat: implement complete Monom class logic

+++ libs/lib_polynom/polynom.h
@@ -39,7 +39,7 @@ class Monom {
     Monom& operator*=(double value);
     Monom& operator/=(double value);
 
-    double calculate(int x, int y, int z) const;
+    double calculate(double x, double y, double z) const;
 
     friend std::ostream& operator<<(std::ostream& os, const Monom& monom);
     friend std::istream& operator>>(std::istream& is, Monom& monom);
